+++ .env.example
@@ -1,12 +1,5 @@
-# General
+LOG_LEVEL=WARNING
+# This directory will be used to save downloaded file
 WORKDIR=./workdir
-LOG_LEVEL=DEBUG
-
 # MediaGrabber runs long tasks and need increased heartbeat and connection timeout
 AMQP_URL=amqp://guest@guest:5672/vhost?heartbeat=300&blocked_connection_timeout=300&socket_timeout=300&stack_timeout=300
-
-# AWS Settings
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
-AWS_DEFAULT_REGION=
-AWS_BUCKET=
