+++ .vscode/launch.json
@@ -12,14 +12,22 @@
             "console": "integratedTerminal"
         },
         {
-            "name": "CLI grab",
+            "name": "CLI Retrieve",
             "type": "python",
             "request": "launch",
-            "program": "./mediagrabber/cli.py",
+            "program": "./app/cli/cli.py",
             "args": [
-                "grab",
+                "retrieve",
                 "https://abcnews.go.com/Technology/video/california-judge-orders-uber-lyft-reclassify-drivers-employees-72302309",
             ],
+            "console": "integratedTerminal"
+        },
+        {
+            "name": "AMQP Application",
+            "type": "python",
+            "request": "launch",
+            "program": "./app/amqp/amqp.py",
+            "args": [],
             "console": "integratedTerminal"
         }
     ]
