+++ app/amqp/amqp.py
@@ -8,9 +8,23 @@ from mediagrabber.config import Config
 from app.amqp.consumer import Consumer
 
 
-def grab(service: MediaGrabber, payload: dict) -> dict:
-    urls = service.grab(payload["url"])
+def retrieve(service: MediaGrabber, payload: dict) -> dict:
+    urls = service.retrieve(payload["url"])
     return {"urls": urls}
+
+class MessageProcessor(object):
+    def __init__(self, service: MediaGrabber, callback):
+        self.service = service
+        self.callback = callback
+
+    # Returns dict when message is processed (successfully or with errors).
+    def process(self, payload: dict) -> Optional[dict]:
+        try:
+            response = self.callback(self.service, payload)
+            return {**payload, **response, "success": True}
+        except MediaGrabberError as err:
+            logging.error(err)
+            return {**payload, "success": False, "error": err.__dict__}
 
 
 if __name__ == "__main__":
@@ -26,9 +40,9 @@ if __name__ == "__main__":
     # We process a heavy tasks, don't need to prefetch more than one message
     channel.basic_qos(prefetch_count=1)
 
-    # Create two consumers for memorize and recognize queues
+    # Create consumer
     service: MediaGrabber = injector.get(MediaGrabber)
-    Consumer(service, channel, Config.queue_in(), Config.queue_out(), grab)
+    Consumer(service, channel, Config.queue_in(), Config.queue_out(), retrieve)
 
     try:
         channel.start_consuming()
