+++ app/amqp/consumer.py
@@ -1,48 +1,48 @@
+from app.amqp.amqp import MessageProcessor
 import json
-from typing import Optional
+from abc import ABC, abstractmethod
+from typing import List
 from pika.adapters.blocking_connection import BlockingChannel
-from mediagrabber.core import MediaGrabber, MediaGrabberError
 import logging
 
 
-class MessageProcessor(object):
-    def __init__(self, service: MediaGrabber, callback):
-        self.service = service
-        self.callback = callback
-
-    # Returns dict when message is processed (successfully or with errors).
-    def process(self, payload: dict) -> Optional[dict]:
-        try:
-            response = self.callback(self.service, payload)
-            return {**payload, **response, "success": True}
-        except MediaGrabberError as err:
-            logging.error(err)
-            return {**payload, "success": False, "error": err.__dict__}
+class MessageProcessorInterface(ABC):
+    @abstractmethod
+    def process(self, payload: dict) -> List[dict]:
+        raise NotImplementedError
 
 
 class Consumer(object):
+    """Consumer declares Input Queue and Output Queue, and starts listen the Input Queue.
+    Each received message will be passed to the Processor and Processor must return the iterable
+    set of messages. Each will be published to the Output Queue and only after that incoming
+    message from the Input Queue will be acknowledged.
+    """
+
     def __init__(
         self,
-        service: MediaGrabber,
         channel: BlockingChannel,
         queue_in: str,
         queue_out: str,
-        callback,
+        processor: MessageProcessorInterface,
     ):
         self.channel = channel
         self.queue_out = queue_out
         self.channel.queue_declare(queue_in, durable=True)
         self.channel.queue_declare(queue_out, durable=True)
+        self.processor = processor
         self.channel.basic_consume(queue_in, self.on_message)
-        self.processor = MessageProcessor(service, callback)
 
     def on_message(self, ch: BlockingChannel, method, properties, body: str):
+        # We received message from the incoming queue
         logging.info(f"Incoming message received: {body}")
-
         payload = json.loads(body)
-        payload = self.processor.process(payload)
-        body = json.dumps(payload)
 
-        logging.info(f"Outcoming message prepared: {body}")
-        self.channel.basic_publish("", self.queue_out, body)
+        # We process this message through processor
+        for message in self.processor.process(payload):
+            body = json.dumps(payload)
+            logging.info(f"Outcoming message prepared: {body}")
+            self.channel.basic_publish("", self.queue_out, body)
+
+        # We have processed all the messages from the processor, now we ack incoming message
         ch.basic_ack(delivery_tag=method.delivery_tag)
